feat(MarketOffsets) : support the 64-bit MW2 build
The September 2026 Steam update replaced iw4mp.exe with an AMD64 build.
That broke every hard-coded address (new image base, ASLR enabled), the
4-byte pointer reads and the dvar_t value offset.

- Trainer: 64-bit addresses, ReadPointer (8 bytes), Is64BitProcess and
  ModuleBase helpers.
- Offset interfaces and 32-bit offset classes switched from int to long.
- New iw4mp64 offset class with RVA offsets and dvar = 0x10.
- COD detects the target bitness and selects iw4mp64 when needed.
- FpsUnlocker resolves dvar pointers through module base + RVA.
- External console reports itself unavailable on 64-bit builds (its
  remote-thread stub is x86 only).
- App project builds as x64 so it can hold 64-bit addresses.

Fixes #17

+++ LINDRA - Market/MarketOffsets/Offsets/BlackOpsMP.cs
@@ -9,23 +9,23 @@ namespace MarketOffsets.Offsets
 {
     internal class BlackOpsMP : IFovAndFps, IVisuals, IDisable, IConsole
     {
-        public int dvar { get { return 0x18; } }
-        public int cg_fov { get { return 0xD52624; } }
-        public int cg_fovMin { get { return 0xCC19CC; } }
+        public long dvar { get { return 0x18; } }
+        public long cg_fov { get { return 0xD52624; } }
+        public long cg_fovMin { get { return 0xCC19CC; } }
 
-        public int cg_fovScale { get { return 0xD52388; } }
-        public int com_maxfps { get { return 0X35A039C; } }
+        public long cg_fovScale { get { return 0xD52388; } }
+        public long com_maxfps { get { return 0X35A039C; } }
 
-        public int r_lightMap { get { return 0x48A4CAC; } }
-        public int r_specularMap { get { return 0x48A4C14; } }
-        public int r_filmUseTweaks { get { return 0x48A4A3C; } }
+        public long r_lightMap { get { return 0x48A4CAC; } }
+        public long r_specularMap { get { return 0x48A4C14; } }
+        public long r_filmUseTweaks { get { return 0x48A4A3C; } }
 
-        public int cg_brass { get { return 0xCC1A14; } }
-        public int r_fog { get { return 0x48A4F6C; } }
-        public int r_glow { get { return 0x0; } }
-        public int r_detail { get { return 0x0; } }
-        public int r_detailMap { get { return 0x0; } }
-        public int cbuf_addtext { get { return 0x56EF70; } }
-        public int nop_address { get { return 0x8B5A37; } }
+        public long cg_brass { get { return 0xCC1A14; } }
+        public long r_fog { get { return 0x48A4F6C; } }
+        public long r_glow { get { return 0x0; } }
+        public long r_detail { get { return 0x0; } }
+        public long r_detailMap { get { return 0x0; } }
+        public long cbuf_addtext { get { return 0x56EF70; } }
+        public long nop_address { get { return 0x8B5A37; } }
     }
 }
